backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=13.0
faster-whisper>=1.1.0
deepl>=1.18.0
yt-dlp>=2024.12.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-multipart>=0.0.18
aiofiles>=24.1.0
google-generativeai>=0.8.0
